MB-49454: Protect HashTable::getRandomKey against HashTable::resize
Restructure getRandomKey so that it is a bit more testable and add
code so that the HashTable size is compared against the expected
size after obtaining a HashBucketLock. If the size is not as
expected, getRandomKey must "reconfigure" and effectively start
again.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165550
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/kv_bucket.h
@@ -1109,7 +1109,7 @@ protected:
      * Allows us to override the random function.  This is used for testing
      * purposes where we want a constant number as opposed to a random one.
      */
-    std::function<long()> getRandom = std::rand;
+    std::function<int()> getRandom = std::rand;
 
     /**
      * The Minimum Durability Level enforced by this Bucket:
